% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BasicCpp.R
\name{addBase}
\alias{addBase}
\title{addBase}
\usage{
addBase(x, y)
}
\arguments{
\item{x}{A vector}

\item{y}{A Vector}
}
\value{
Addition of x and y
}
\description{
Add two vectors
}
\examples{
addBase(1:10, 1:10)
}
\author{
Rick Qiu
}
